+++ CHANGELOG.md
@@ -1,4 +1,8 @@
 ## 0.78.0 (Unreleased)
+FEATURES:
+* greenplum: add `maintenance_window` attribute to resource and data source
+* greenplum: support for changing the user's password
+* greenplum: support for changing segment and master resources
 
 ## 0.77.0 (July 27, 2022)
 BUG FIXES: